[PC-8818] Move offerer creation to pages/Offerers/OffererCreation/
