+++ src/repository/siren/__specs__/getSirenInformation.spec.js
